rm game from navbar + move scores views in score app + rm useless files
