+++ services/backend/project/project/urls.py
@@ -24,5 +24,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('user/', include('authentication.urls')),
     path('tournament/', include('tournament.urls')),
+    path('score/', include('score.urls')),
 ]
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
