+++ services/frontend/src/components/morpion/morpion.js
@@ -22,7 +22,7 @@ let onClickHandler = (e) => {
 let GetParties = async (user) => {
 	try {
 		const authtoken = sessionStorage.getItem('authtoken');
-		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/user/get_parties/', {
+		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/score/get_parties/', {
 			method: 'GET',
 			headers: {
 				'Content-Type': 'application/json',
@@ -83,7 +83,7 @@ let GetScore = async () => {
 	// const { t }	= useTranslation();
 	try {
 		const authtoken = sessionStorage.getItem('authtoken');
-		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/user/get_top_score/', {
+		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/score/get_top_score/', {
 			method: 'GET',
 			headers: {
 				'Content-Type': 'application/json',
@@ -311,7 +311,7 @@ const sendScore = async (winner, points) => {
 	const data = { winner: winner, points: points};
 	try {
 
-		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/user/update_score/', {
+		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/score/update_score/', {
 			method: 'POST',
 			headers: {
 				'Content-Type': 'application/json',
@@ -332,7 +332,7 @@ const sendParty = async (winner, points, oponent) => {
 	const data = { winner: winner, points: points, oponent: oponent};
 	try {
 
-		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/user/update_parties/', {
+		const response = await fetch('http://' + window.location.host.split(':')[0] + ':8000/score/update_parties/', {
 			method: 'POST',
 			headers: {
 				'Content-Type': 'application/json',
