+++ services/frontend/src/components/navbar.js
@@ -22,9 +22,9 @@ function NavBar(props) {
                 <Navbar.Toggle aria-controls="navbar" />
                 <Navbar.Collapse id="navbar">
                     <Nav id="navBar" className="me-auto">
-                        <Nav.Link as={Link} to="/Game">{t('Game')}</Nav.Link>
+                        {/* <Nav.Link as={Link} to="/Game">{t('Game')}</Nav.Link> */}
                         <Nav.Link as={Link} to="/tournament">{t('Tournament')}</Nav.Link>
-                        <Nav.Link as={Link} to="/ThreejsGame">{t('3D-Game')}</Nav.Link>
+                        {/* <Nav.Link as={Link} to="/ThreejsGame">{t('3D-Game')}</Nav.Link> */}
                         <Nav.Link as={Link} to="/Morpion">Morpion</Nav.Link>
                         {props.isLoggedIn ? (
                             <>
